fix(rbac): move admin creation to the admin-permissions directory
