+++ plugins/rbac-backend/src/service/enforcer-delegate.ts
@@ -3,11 +3,11 @@ import { Knex } from 'knex';
 
 import EventEmitter from 'events';
 
+import { ADMIN_ROLE_NAME } from '../admin-permissions/admin-creation';
 import {
   RoleMetadataDao,
   RoleMetadataStorage,
 } from '../database/role-metadata';
-import { ADMIN_ROLE_NAME } from '../file-permissions/admin-creation';
 import { mergeRoleMetadata, policiesToString, policyToString } from '../helper';
 import { MODEL } from './permission-model';
